fix(paper): 修 ruff RUF046 + I001 两处 CI 失败
- compose.py:239 RUF046: int(round(float(x))) 多余的 int()
  Python 3 中 round(x) 单参数返回 int，外层 int() 是 no-op。
- test_compose.py I001: imports 段尾的两空行收成一空行（ruff isort 默认）。

+++ services/paper/src/inalpha_paper/strategies/compose.py
@@ -236,7 +236,7 @@ def _compose_buy_hold(
 def _coerce_int(value: Any, lo: int, hi: int, *, default: int) -> int:
     """把 LLM 给的任意值转 int 并 clip 到 [lo, hi]。失败用 default。"""
     try:
-        x = int(round(float(value))) if value is not None else default
+        x = round(float(value)) if value is not None else default
     except (TypeError, ValueError):
         x = default
     return max(lo, min(hi, x))
